+++ src/test/java/uk/ac/standrews/cs/utilities/mi_file/MIFile2DTest.java
@@ -45,7 +45,7 @@ public class MIFile2DTest {
 
     private Set<Point> createReferenceObjects() {
         Set<Point> pts = new HashSet<>();
-        for (double step = 10.0F; step < 80.0F; step+=10.0F) {
+        for (float step = 10.0F; step < 80.0F; step+=10.0F) {
 
             pts.add(new Point(+step + 1.0f, +step + 2.0f ));
             pts.add(new Point(-step - 3.0f , -step - 4.0f ));
@@ -129,14 +129,14 @@ public class MIFile2DTest {
     public void addLinear15() throws Exception {
 
         for (int i = 0; i < 15; i++) {
-            file.add(new Point((double) i, 0.0F));
+            file.add(new Point( i, 0.0F));
         }
 
         assertEquals(15, file.size());
 
 
         for (int i = 0; i < 15; i++) {
-            Point p = new Point((double) i, 0.0F);
+            Point p = new Point( i, 0.0F);
             assertTrue(file.contains(p));
         }
     }
@@ -151,29 +151,29 @@ public class MIFile2DTest {
 
         // lay down 20 points in a line
         for (int i = 0; i < 3; i++) {
-            file.add(new Point((double) i * 10, 0.0F));
+            file.add(new Point( i * 10, 0.0F));
         }
         // new lay down 20 points in a line - that are all close (4 away) to the first 20
         for (int i = 0; i < 3; i++) {
-            file.add(new Point((double) (i * 10) + 4.0F, 0.0F));
+            file.add(new Point((i * 10) + 4.0F, 0.0F));
         }
         // new lay down another 20 points in a line - that are all close (1 away) to the second 20
         for (int i = 0; i < 3; i++) {
-            file.add(new Point((double) (i * 10) + 4.5F, 0.0F));
+            file.add(new Point( (i * 10) + 4.5F, 0.0F));
         }
 
         assertEquals(9, file.size());
 
         for (int i = 0; i < 3; i++) {
-            assertTrue(file.contains(new Point((double) i * 10, 0.0F)));
+            assertTrue(file.contains(new Point( i * 10, 0.0F)));
         }
         // new lay down 20 points in a line - that are all close (4 away) to the first 20
         for (int i = 0; i < 3; i++) {
-            assertTrue(file.contains(new Point((double) (i * 10) + 4.0F, 0.0F)));
+            assertTrue(file.contains(new Point( (i * 10) + 4.0F, 0.0F)));
         }
         // new lay down another 20 points in a line - that are all close (1 away) to the second 20
         for (int i = 0; i < 3; i++) {
-            assertTrue(file.contains(new Point((double) (i * 10) + 4.5F, 0.0F)));
+            assertTrue(file.contains(new Point( (i * 10) + 4.5F, 0.0F)));
         }
     }
 
@@ -185,7 +185,7 @@ public class MIFile2DTest {
 
         int count = 0;
 
-        for (double step = 1.0F; step < 50.0F; step++) {
+        for (float step = 1.0F; step < 50.0F; step++) {
 
             file.add(new Point(+step, +step));
             count++;
@@ -210,7 +210,7 @@ public class MIFile2DTest {
         int count = addSquares();
         assertEquals(count, file.size());
 
-        for (double step = 1.0F; step < 50.0F; step++) {
+        for (float step = 1.0F; step < 50.0F; step++) {
 
             assertTrue(file.contains(new Point(+step, +step)));
             assertTrue(file.contains(new Point(-step, +step)));
